Fix overflows in AnimatedContentSamplerTest's custom RNG
AnimatedContentSamplerTest eschewed the RNG provided in //base and
implemented its own for performance. In doing so, it didn't get overflow
correct. Code using signed integers is required to never overflow in
C++.

Instead, run the RNG over uint32_t and cast to int at the end.

Bug: 1394755
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4994874
Commit-Queue: David Benjamin <davidben@chromium.org>
Auto-Submit: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218373}

+++ media/capture/content/animated_content_sampler_unittest.cc
@@ -38,7 +38,7 @@ class AnimatedContentSamplerTest : public ::testing::Test {
   ~AnimatedContentSamplerTest() override = default;
 
   void SetUp() override {
-    rand_seed_ = static_cast<int>(
+    rand_seed_ = static_cast<uint32_t>(
         (InitialTestTimeTicks() - base::TimeTicks()).InMicroseconds());
     sampler_ = std::make_unique<AnimatedContentSampler>(GetMinCapturePeriod());
   }
@@ -90,10 +90,10 @@ class AnimatedContentSamplerTest : public ::testing::Test {
   // builds.  The following is a very simple, deterministic LCG:
   int NextRandomInt() {
     rand_seed_ = (1103515245 * rand_seed_ + 12345) % (1 << 31);
-    return rand_seed_;
+    return static_cast<int>(rand_seed_);
   }
 
-  int rand_seed_;
+  uint32_t rand_seed_;
   std::unique_ptr<AnimatedContentSampler> sampler_;
 };
